# backend/backend-api/src/main/resources/static/openapi.yml
openapi: "3.1.0"

info:
  title: BAS Verification API
  version: "1.0.0"

servers:
  - url: http://localhost:8080

security:
  - bearerAuth: []

paths:
  /api/auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Регистрация
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Пользователь создан
        "400":
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Вход
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Неверный email или пароль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/routes:
    post:
      tags: [Routes]
      summary: Создать маршрут и отправить на проверку
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRouteRequest"
      responses:
        "201":
          description: Маршрут создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"
        "400":
          description: Менее 2 точек
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      tags: [Routes]
      summary: Список своих маршрутов
      responses:
        "200":
          description: Список маршрутов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/routes/{id}:
    get:
      tags: [Routes]
      summary: Маршрут по id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Маршрут
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Не найден или чужой

  /api/zones:
    post:
      tags: [Zones]
      summary: Создать зону
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateZoneRequest"
      responses:
        "201":
          description: Зона создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [Zones]
      summary: Список всех зон
      responses:
        "200":
          description: Список зон
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ZoneResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /api/admin/routes:
    get:
      tags: [Admin]
      summary: Список всех заявок
      responses:
        "200":
          description: Список всех заявок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /api/admin/routes/pending:
    get:
      tags: [Admin]
      summary: Заявки на ручное рассмотрение
      responses:
        "200":
          description: Список заявок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /api/admin/routes/{id}/approve:
    post:
      tags: [Admin]
      summary: Согласовать маршрут
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Маршрут согласован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"
        "400":
          description: Маршрут уже обработан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Маршрут не найден

  /api/admin/routes/{id}/reject:
    post:
      tags: [Admin]
      summary: Отклонить маршрут
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Маршрут отклонён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResponse"
        "400":
          description: Маршрут уже обработан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Маршрут не найден

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    Unauthorized:
      description: Неавторизован или токен истёк
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:

    RegisterRequest:
      type: object
      required: [fullName, email, password]
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]

    CreateRouteRequest:
      type: object
      required: [name, points]
      properties:
        name:
          type: string
        points:
          type: array
          minItems: 2
          items:
            $ref: "#/components/schemas/Coordinate"

    Coordinate:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number

    RouteResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        userFullName:
          type: string
        status:
          type: string
          enum: [SUBMITTED, APPROVED, REJECTED]
        autoCheckResult:
          type: boolean
        createdAt:
          type: string
          format: date-time
        points:
          type: array
          items:
            $ref: "#/components/schemas/Coordinate"

    CreateZoneRequest:
      type: object
      required: [name, coordinates]
      properties:
        name:
          type: string
        coordinates:
          type: array
          minItems: 3
          items:
            $ref: "#/components/schemas/Coordinate"
          description: Вершины полигона, первая и последняя совпадают

    ZoneResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        coordinates:
          type: array
          items:
            $ref: "#/components/schemas/Coordinate"
        createdAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string